+++ src/index.js
@@ -3,14 +3,16 @@
 const useScale = require('./rules/use-scale');
 const preferToken = require('./rules/prefer-token');
 const noOffscaleTransform = require('./rules/no-offscale-transform');
+const useMotionScale = require('./rules/use-motion-scale');
 
-const rules = [useScale, preferToken, noOffscaleTransform];
+const rules = [useScale, preferToken, noOffscaleTransform, useMotionScale];
 
 module.exports = rules;
 module.exports.rules = {
   [useScale.ruleName]: useScale,
   [preferToken.ruleName]: preferToken,
   [noOffscaleTransform.ruleName]: noOffscaleTransform,
+  [useMotionScale.ruleName]: useMotionScale,
 };
 module.exports.configs = {
   recommended: require('./configs/recommended'),
@@ -19,6 +21,7 @@ module.exports.configs = {
   expanded: require('./configs/expanded'),
   logical: require('./configs/logical'),
   migration: require('./configs/migration'),
+  motion: require('./configs/motion'),
 };
 module.exports.eslint = require('./eslint');
 module.exports.presets = require('./presets');